freddy_bot/requirements.txt
flask>=3.0
pymongo>=4.7
requests>=2.32
python-dotenv>=1.0
gunicorn>=22.0
